add separate view for walls images
